requirements.txt
python-telegram-bot[job-queue]==20.8
yt-dlp>=2026.3.17
requests>=2.32.0
python-dotenv>=1.0.0
